+++ index.html
@@ -2,6 +2,7 @@
 <html lang="es">
     <head>
         <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link
             href="https://fonts.googleapis.com/css2?family=Caveat&family=M+PLUS+1+Code:wght@100&family=Marvel&display=swap"
             rel="stylesheet"
